// WADNR.Web/src/app/shared/generated/model/focus-area-upsert-request.ts
/**
 * WADNR.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class FocusAreaUpsertRequest { 
    FocusAreaName?: string | null;
    FocusAreaStatusID?: number;
    DNRUplandRegionID?: number;
    PlannedFootprintAcres?: number | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface FocusAreaUpsertRequestForm { 
    FocusAreaName?: FormControl<string>;
    FocusAreaStatusID?: FormControl<number>;
    DNRUplandRegionID?: FormControl<number>;
    PlannedFootprintAcres?: FormControl<number>;
}

export class FocusAreaUpsertRequestFormControls { 
    public static FocusAreaName = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static FocusAreaStatusID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static DNRUplandRegionID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static PlannedFootprintAcres = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}